// src/app/models/lender.ts
export interface Lender {
    lenderId: string;
    legalName: string;
    legalIdType: string;
    legalIdNumber: string;
    contactPersonnel: string;
    jobTitle: string;
    contactNumber: string;
    email: string;
    line1: string;
    line2: string;
    line3: string;
    state: string;
    zipCode: string;
}